+++ CHANGELOG.md
@@ -28,9 +28,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /api/1.5/letsencrypt/autorenew `POST`
   - /api/1.5/letsencrypt/dnsrecords `GET`
   - /api/2.0/vault/ping `GET`
+  - /api/2.0/vault/bucket/:bucket/key/:key/values `GET`
+  - /api/2.0/servercheck `GET`
 
 ### Changed
 - Fix to traffic_ops_ort.pl to strip specific comment lines before checking if a file has changed.  Also promoted a changed file message from DEBUG to ERROR for report mode.
+- Fixed Traffic Portal regenerating CDN DNSSEC keys with the wrong effective date
 - Updated The Traffic Ops Python, Go and Java clients to use API version 2.0 (when possible)
 - Updated CDN-in-a-Box scripts and enroller to use TO API version 2.0
 - Updated numerous, miscellaneous tools to use TO API version 2.0
@@ -52,8 +55,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /cdns/:id (GET)
   - /cdns/name/:name (GET)
   - /cdns/usage/overview
+  - /deliveryservice_matches
   - /deliveryservice_user
   - /deliveryservice_user/:dsId/:userId
+  - /deliveryservices/hostname/:name/sslkeys
   - /deliveryservices/{dsid}/regexes/{regexid} (GET)
   - /deliveryservices/:id (GET)
   - /deliveryservices/:id/state
@@ -62,12 +67,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /divisions/name/:name
   - /hwinfo/dtdata
   - /jobs/:id
+  - /keys/ping
   - /logs/:days/days
   - /parameters/:id/profiles
   - /parameters/:id/unassigned_profiles
+  - /parameters/profile/:name
   - /parameters/validate
+  - /profile/:id/unassigned_parameters
   - /regions/:region_name/phys_locations
   - /regions/name/:region_name
+  - /riak/bucket/:bucket/key/:key/vault
   - /riak/ping
   - /riak/stats
   - /servercheck/aadata
@@ -77,6 +86,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /traffic_monitor/stats
   - /types/trimmed
   - /user/current/jobs
+  - /servers/checks
 
 ## [4.0.0] - 2019-12-16
 ### Added